1.0.0

+++ controllers/CategoryController.php
@@ -2,7 +2,8 @@
 
 namespace app\controllers;
 
-use yii;
+use app\models\Tickets;
+use app\models\Documents;
 use app\models\Categories;
 use app\models\CategorySearch;
 
@@ -10,20 +11,20 @@ use app\models\CategorySearch;
  * CategoryController implements the CRUD actions for Categories model.
  */
 class CategoryController extends AppController{
+    
     /**
      * @inheritDoc
      */
-    public function behaviors()
-    {
+    public function behaviors() : array{
         return array_merge(
             parent::behaviors(),
             [
                 'verbs' => [
-                    'class' => yii\filters\VerbFilter::class,
+                    'class' => \yii\filters\VerbFilter::class,
                     'actions' => [
-                        'delete' => ['POST'],
-                    ],
-                ],
+                        'delete' => ['POST']
+                    ]
+                ]
             ]
         );
     }
@@ -33,13 +34,12 @@ class CategoryController extends AppController{
      *
      * @return string
      */
-    public function actionIndex(){
+    public function actionIndex() : string{
         $searchModel = new CategorySearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
+        $dataProvider = $searchModel->search(\yii::$app->request->get());
         return $this->render('index', [
             'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
+            'dataProvider' => $dataProvider
         ]);
     }
 
@@ -47,32 +47,31 @@ class CategoryController extends AppController{
      * Displays a single Categories model.
      * @param int $id ID
      * @return string
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionView($id){
+    public function actionView(int $id) : string{
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $this->findModel($id)
         ]);
     }
 
     /**
      * Creates a new Categories model.
      * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|yii\web\Response
+     * @return string|\yii\web\Response
      */
-    public function actionCreate(){
+    public function actionCreate() : string|\yii\web\Response{
         $model = new Categories();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
+        if (\Yii::$app->request->isPost){
+            if ($model->load(\yii::$app->request->post()) && $model->save()) {
                 return $this->redirect(['view', 'id' => $model->id]);
             }
-        } else {
+        }
+        else{
             $model->loadDefaultValues();
         }
-
         return $this->render('create', [
-            'model' => $model,
+            'model' => $model
         ]);
     }
 
@@ -80,18 +79,16 @@ class CategoryController extends AppController{
      * Updates an existing Categories model.
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param int $id ID
-     * @return string|yii\web\Response
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @return string|\yii\web\Response
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionUpdate($id){
+    public function actionUpdate(int $id) : string|\yii\web\Response{
         $model = $this->findModel($id);
-
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+        if(\Yii::$app->request->isPost && $model->load(\yii::$app->request->post()) && $model->save()){
             return $this->redirect(['view', 'id' => $model->id]);
         }
-
         return $this->render('update', [
-            'model' => $model,
+            'model' => $model
         ]);
     }
 
@@ -99,12 +96,21 @@ class CategoryController extends AppController{
      * Deletes an existing Categories model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param int $id ID
-     * @return yii\web\Response
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @return \yii\web\Response
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionDelete($id){
-        $this->findModel($id)->delete();
-
+    public function actionDelete(int $id) : \yii\web\Response{
+        if(Tickets::find()->where(['category_id' => $id])->count() === 0){
+            if(Documents::find()->where(['category_id' => $id])->count() === 0){
+                $this->findModel($id)->delete();
+            }
+            else{
+                \Yii::$app->getSession()->addFlash('error', 'Данная категория еще используется! '. \yii\helpers\Html::a(\yii\helpers\Html::encode('Перейти к данным документам'), \yii\helpers\Url::to(['documents/', 'DocumentSearch[category_id]' => $id]), ['class' => 'link-dark link-offset-2 link-underline-opacity-50 link-underline-opacity-100-hover', 'title' => 'Перейти']));
+            }
+        }
+        else{
+            \Yii::$app->getSession()->addFlash('error', 'Данная категория еще используется! '. \yii\helpers\Html::a(\yii\helpers\Html::encode('Перейти к данным обращениям'), \yii\helpers\Url::to(['tickets/', 'TicketSearch[category_id]' => $id]), ['class' => 'link-dark link-offset-2 link-underline-opacity-50 link-underline-opacity-100-hover', 'title' => 'Перейти']));
+        }
         return $this->redirect(['index']);
     }
 
@@ -113,13 +119,12 @@ class CategoryController extends AppController{
      * If the model is not found, a 404 HTTP exception will be thrown.
      * @param int $id ID
      * @return Categories the loaded model
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id){
-        if (($model = Categories::findOne(['id' => $id])) !== null) {
+    protected function findModel(int $id) : Categories{
+        if(($model = Categories::findOne(['id' => $id])) !== null){
             return $model;
         }
-
-        throw new yii\web\NotFoundHttpException('The requested page does not exist.');
+        throw new \yii\web\NotFoundHttpException('Страница не найдена.');
     }
 }
