+++ controllers/CityController.php
@@ -2,27 +2,28 @@
 
 namespace app\controllers;
 
-use yii;
 use app\models\Cities;
+use app\models\Tickets;
 use app\models\CitySearch;
 
 /**
  * CitiyController implements the CRUD actions for Cities model.
  */
 class CityController extends AppController{
+
     /**
      * @inheritDoc
      */
-    public function behaviors(){
+    public function behaviors() : array{
         return array_merge(
             parent::behaviors(),
             [
                 'verbs' => [
-                    'class' => yii\filters\VerbFilter::class,
+                    'class' => \yii\filters\VerbFilter::class,
                     'actions' => [
-                        'delete' => ['POST'],
-                    ],
-                ],
+                        'delete' => ['POST']
+                    ]
+                ]
             ]
         );
     }
@@ -32,12 +33,12 @@ class CityController extends AppController{
      *
      * @return string
      */
-    public function actionIndex(){
+    public function actionIndex() : string{
         $searchModel = new CitySearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        $dataProvider = $searchModel->search(\yii::$app->request->get());
         return $this->render('index', [
             'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
+            'dataProvider' => $dataProvider
         ]);
     }
 
@@ -45,32 +46,31 @@ class CityController extends AppController{
      * Displays a single Cities model.
      * @param int $id ID
      * @return string
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionView($id){
+    public function actionView(int $id) : string{
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $this->findModel($id)
         ]);
     }
 
     /**
      * Creates a new Cities model.
      * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|yii\web\Response
+     * @return string|\yii\web\Response
      */
-    public function actionCreate(){
+    public function actionCreate() : string|\yii\web\Response{
         $model = new Cities();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
+        if(\Yii::$app->request->isPost){
+            if($model->load(\yii::$app->request->post()) && $model->save()) {
                 return $this->redirect(['view', 'id' => $model->id]);
             }
-        } else {
+        }
+        else{
             $model->loadDefaultValues();
         }
-
         return $this->render('create', [
-            'model' => $model,
+            'model' => $model
         ]);
     }
 
@@ -78,18 +78,16 @@ class CityController extends AppController{
      * Updates an existing Cities model.
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param int $id ID
-     * @return string|yii\web\Response
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @return string|\yii\web\Response
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionUpdate($id){
+    public function actionUpdate(int $id) : string|\yii\web\Response{
         $model = $this->findModel($id);
-
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+        if(\Yii::$app->request->isPost && $model->load(\yii::$app->request->post()) && $model->save()){
             return $this->redirect(['view', 'id' => $model->id]);
         }
-
         return $this->render('update', [
-            'model' => $model,
+            'model' => $model
         ]);
     }
 
@@ -97,12 +95,16 @@ class CityController extends AppController{
      * Deletes an existing Cities model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param int $id ID
-     * @return yii\web\Response
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @return \yii\web\Response
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionDelete($id){
-        $this->findModel($id)->delete();
-
+    public function actionDelete(int $id) : string|\yii\web\Response{
+        if(Tickets::find()->where(['city_id' => $id])->count() === 0){
+            $this->findModel($id)->delete();
+        }
+        else{
+            \Yii::$app->getSession()->addFlash('error', 'Данный город еще используется! '. \yii\helpers\Html::a(\yii\helpers\Html::encode('Перейти к данным обращениям'), \yii\helpers\Url::to(['tickets/', 'TicketSearch[city_id]' => $id]), ['class' => 'link-dark link-offset-2 link-underline-opacity-50 link-underline-opacity-100-hover', 'title' => 'Перейти']));
+        }
         return $this->redirect(['index']);
     }
 
@@ -111,12 +113,12 @@ class CityController extends AppController{
      * If the model is not found, a 404 HTTP exception will be thrown.
      * @param int $id ID
      * @return Cities the loaded model
-     * @throws yii\web\NotFoundHttpException if the model cannot be found
+     * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id){
+    protected function findModel(int $id) : Cities{
         if (($model = Cities::findOne(['id' => $id])) !== null) {
             return $model;
         }
-        throw new yii\web\NotFoundHttpException('The requested page does not exist.');
+        throw new \yii\web\NotFoundHttpException('Страница не найдена.');
     }
 }
