+++ controllers/TicketController.php
@@ -16,7 +16,7 @@ class TicketController extends AppController{
     /**
     * @inheritDoc
     */
-   public function behaviors(){
+   public function behaviors() : array{
        return array_merge(
            parent::behaviors(),
            [
@@ -35,12 +35,12 @@ class TicketController extends AppController{
      *
      * @return string
      */
-    public function actionIndex(){
+    public function actionIndex() : string{
         $searchModel = new TicketSearch();
         $dataProvider = $searchModel->search(\yii::$app->request->get());
         return $this->render('index', [
             'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
+            'dataProvider' => $dataProvider
         ]);
     }
 
@@ -50,7 +50,7 @@ class TicketController extends AppController{
      * @return string
      * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionView($id){
+    public function actionView(int $id) : string{
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -61,19 +61,19 @@ class TicketController extends AppController{
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return string|\yii\web\Response
      */
-    public function actionCreate(){
+    public function actionCreate() : string|\yii\web\Response{
         $model = new Tickets();
         if(\Yii::$app->request->isAjax && $model->load(\Yii::$app->request->post())){
             \Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
             return \yii\widgets\ActiveForm::validate($model);
         }
-        if($model->load($this->request->post())){
+        if($model->load(\yii::$app->request->post())){
             if($model->validate()){
                 if($model->save()){
                     return $this->redirect(['view', 'id' => $model->id]);
                 }
                 else{
-                    \Yii::$app->session->setFlash('error', 'Произошла ошибка при создании обращения');                  
+                    \Yii::$app->session->addFlash('error', 'Произошла ошибка при создании обращения');                  
                     return $this->render('create', [
                         'model' => $model,
                         'categories' => new Categories()
@@ -105,13 +105,13 @@ class TicketController extends AppController{
      * @return string|\yii\web\Response
      * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    public function actionUpdate($id){
+    public function actionUpdate(int $id) : string|\yii\web\Response{
         $model = $this->findModel($id);
         if(\Yii::$app->request->isAjax && $model->load(\Yii::$app->request->post())){
             \Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
             return \yii\widgets\ActiveForm::validate($model);
         }
-        if($model->load($this->request->post())){
+        if($model->load(\yii::$app->request->post())){
             if($model->validate()){
                 if($model->save()){
                     return $this->redirect(['view', 'id' => $model->id]);
@@ -139,11 +139,16 @@ class TicketController extends AppController{
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param int $id ID
      * @return \yii\web\Response
-     * @throws \yii\web\NotFoundHttpException if the model cannot be found
+     * @throws \yii\web\NotFoundHttpException|\yii\web\ForbiddenHttpException if the model cannot be found
      */
-    public function actionDelete($id){
-        $this->findModel($id)->delete();
-        return $this->redirect(['index']);
+    public function actionDelete(int $id) : \yii\web\Response{
+        if(\Yii::$app->user->can('admin')){
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }
+        else{
+            throw new \yii\web\ForbiddenHttpException('Доступ только у администраторов');
+        }
     }
 
     /**
@@ -153,10 +158,10 @@ class TicketController extends AppController{
      * @return Tickets the loaded model
      * @throws \yii\web\NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id){
+    protected function findModel(int $id) : Tickets{
         if(($model = Tickets::findOne(['id' => $id])) !== null){
             return $model;
         }
-        throw new \yii\web\NotFoundHttpException('The requested page does not exist.');
+        throw new \yii\web\NotFoundHttpException('Страница не найдена.');
     }
 }
